+++ src/index.ts
@@ -34,7 +34,6 @@ export {
 export {
     ArgumentsReader,
     Option,
-    BindOption,
     CommentStyle,
     JSX,
     LogLevel,