+++ src/lib/output/index.ts
@@ -1,4 +1,4 @@
 export { PageEvent, RendererEvent, MarkdownEvent } from "./events";
 export { Renderer } from "./renderer";
-export { Output, Router, type MinimalDocument } from "./output";
+export { Output, type MinimalDocument } from "./output";
 export { DefaultThemeRenderContext } from "./themes/default/DefaultThemeRenderContext";