+++ src/lib/utils/index.ts
@@ -27,7 +27,6 @@ export { DefaultMap } from "./map";
 export {
     ArgumentsReader,
     Option,
-    BindOption,
     CommentStyle,
     Options,
     PackageJsonReader,